refactor(chocolate): use content API, constants, fix C90 warnings
- Replace local dmcp_weapon_available_in_mode() with dmcp_is_item_available()
- Replace local dmcp_spawn_available_in_mode() with dmcp_is_enemy_spawnable()
- Add dmcp_to_gamemode() converter for Chocolate Doom's GameMode_t
- Add dmcp_float_to_fixed() and dmcp_degrees_to_angle() to dmcp_mappings.h
- Use DMCP_PLAYER_MAX_HEALTH, DMCP_ITEM_STIMPACK, etc. constants
- Fix C90 mixed declaration warnings in state_*.c files
- Remove float casts when populating hp/armor (now int32_t)

+++ adapters/chocolate-doom/state_enemies.c
@@ -46,8 +46,8 @@ int dmcp_chocolate_populate_enemies(dmcp_snapshot_t* snap) {
       memset(&enemy, 0, sizeof(enemy));
 
       enemy.id     = count;
-      enemy.hp     = (float)mo->health;
-      enemy.max_hp = (float)mo->info->spawnhealth;
+      enemy.hp     = mo->health;
+      enemy.max_hp = mo->info->spawnhealth;
 
       // Get current position and angle
       enemy.position.x = dmcp_fixed_to_float(mo->x);
